requirements.txt
boto3>=1.9.174
Django>=2.2.3
djangorestframework==3.11.0
django-rest-swagger==2.2.0
docutils>=0.14,<0.16
gunicorn>=19.9.0
jmespath>=0.9.4
python-dateutil>=2.8.0
pytz>=2019.1
s3transfer>=0.2.1
six>=1.12.0
sqlparse>=0.3.0
urllib3>=1.25.3
elasticsearch-dsl>=5.0.0,<6.0.0
whitenoise>=4.1.3
django-cors-headers>=3.2.0
tldextract
